can now communicate wiht the database.
first make in the contorller the command, to it communicates. then say what function it needs to do, this fucntion will link to the services.ts

in here you name the repository functions you want to call. these fucntions are set in the repository.ts. normally you wont have to do this, but the new version wont allow it to be easier/ te repository creates the struct, where you create information, change it, and then put it back in, or send it back in. its that simple. you can't do that in the services. and you cant go straight from controller to repository, ebcause you need to call on what to return. in this case you don't return, but in the future you might.

this is the ORM, and its complicated, but functional.