succesfully made, now it needs to be tested properly
testing first when you log in. will it pass through and get the proper values correct.

then testing the enable page. if the code is succesfully accepted. AND only that code. then do the procedure where it is accepted. then it shoudl change where it will be accepted. then if I logout. it will think it needs to be accepted. then, input the code there. and voila